Remove UserRepository interface
* 미션 1단계 요구사항 구현에 불필요한 부분을 제거하였다
* OnMemoryUserRepository를 UserRepository로 이름 변경
