+++ README.md
@@ -9,16 +9,19 @@
 - [x] 회원 가입 기능 구현
   - 회원 가입 양식: /users/form.html
   - 요청 메서드: POST /users
+  - 전달받은 값으로 User 객체를 생성하여 저장
+    - getter, setter 메소드를 생성한다
+  - 생성된 User 객체는 ArrayList로 구현한 사용자 목록에 저장한다.
   - 회원 가입 후: 유저 목록으로 리다이렉트 (redirect:/users)
 - [x] 회원 조회 기능 구현
   - 요청 메서드: GET /users
   - 회원 목록 뷰 템플릿: /users/list.html
 - [x] 회원 프로필 조회 기능 구현
+  - 회원 목록에서 닉네임을 누르면 프로필 페이지로 이동하도록 한다. 
   - 요청 메서드: GET /users/{userId}
   - 회원 프로필 뷰 템플릿: /users/profile.html
 - [x] 템플릿을 이용하여 HTML 파일들의 중복 제거
 - [x] DB는 사용하지 않는다.
-  - 단, 이후 DB를 사용하도록 변경할 수 있어야 한다.
 
 ### 구현과정
